Reduce Remix UI style browser graph
