+++ packages/ui/.changes/patch.style-ssr-client-boundary-cleanup.md
@@ -1,0 +1,1 @@
+Reduced browser JavaScript for Remix UI style and prop patching by sharing CSS property helpers across already-downloaded modules and keeping SSR-only prop/tag tables in the server renderer.
