+++ packages/ui/src/runtime/core/attributes.ts
@@ -1,4 +1,4 @@
-import { normalizeCssValue } from '../../style/style.ts'
+import { normalizeCssValue, toCssPropertyName } from '../../style/properties.ts'
 import { normalizeSvgAttribute } from '../svg-attributes.ts'
 import type { ElementProps } from '../jsx.ts'
 
@@ -26,29 +26,6 @@ const BOOLEANISH_STRING_ATTRIBUTES = new Set([
   'preserveAlpha',
   'spellcheck',
 ])
-
-export const FRAMEWORK_PROPS = new Set(['children', 'mix', 'key', 'animate', 'innerHTML', 'on'])
-
-export const SELF_CLOSING_TAGS = new Set([
-  'area',
-  'base',
-  'br',
-  'col',
-  'embed',
-  'hr',
-  'img',
-  'input',
-  'link',
-  'meta',
-  'param',
-  'source',
-  'track',
-  'wbr',
-])
-
-export function isChildlessElement(name: string): boolean {
-  return SELF_CLOSING_TAGS.has(name)
-}
 
 export function canUseProperty(
   element: Element,
@@ -85,20 +62,14 @@ export function isBooleanishStringAttribute(name: string): boolean {
   return BOOLEANISH_STRING_ATTRIBUTES.has(name)
 }
 
-export function shouldStringifyBooleanAttribute(name: string): boolean {
-  return isBooleanishStringAttribute(name) || name === 'value'
-}
-
 export function serializeStyleObject(style: Record<string, unknown>): string {
   let parts: string[] = []
 
   for (let [key, value] of Object.entries(style)) {
-    if (value == null) continue
-    if (typeof value === 'boolean') continue
-    if (typeof value === 'number' && !Number.isFinite(value)) continue
+    let cssValue = styleValueToCss(key, value)
+    if (cssValue === undefined) continue
 
-    let cssKey = toKebabCase(key)
-    let cssValue = Array.isArray(value) ? value.join(', ') : normalizeCssValue(key, value)
+    let cssKey = toCssPropertyName(key)
 
     parts.push(`${cssKey}: ${cssValue};`)
   }
@@ -106,13 +77,21 @@ export function serializeStyleObject(style: Record<string, unknown>): string {
   return parts.join(' ')
 }
 
+export function styleValueToCss(name: string, value: unknown): string | undefined {
+  if (value == null) return undefined
+  if (typeof value === 'boolean') return undefined
+  if (typeof value === 'number' && !Number.isFinite(value)) return undefined
+
+  if (Array.isArray(value)) {
+    return value.join(', ')
+  }
+
+  return normalizeCssValue(name, value)
+}
+
 export function getMergedClassName(props: ElementProps): string | undefined {
   let classAttr = typeof props.class === 'string' ? props.class : ''
   let className = typeof props.className === 'string' ? props.className : ''
   let merged = classAttr && className ? `${classAttr} ${className}` : classAttr || className
   return merged || undefined
-}
-
-export function toKebabCase(value: string): string {
-  return value.replace(/[A-Z]/g, (match) => `-${match.toLowerCase()}`)
 }
