+++ packages/ui/src/runtime/scheduler.ts
@@ -8,7 +8,7 @@ import {
   setActiveSchedulerUpdateParents,
 } from './reconcile.ts'
 import { defaultStyleManager } from './diff-props.ts'
-import type { StyleManager } from '../style/index.ts'
+import type { StyleManager } from '../style/stylesheet.ts'
 
 type EmptyFn = () => void
 type SchedulerPhaseType = 'beforeUpdate' | 'commit'
