+++ packages/ui/src/style/css-mixin.ts
@@ -3,8 +3,8 @@ import type { MixinFactory } from '../runtime/mixins/mixin.ts'
 import type { MixinDescriptor } from '../runtime/mixins/mixin.ts'
 import type { ElementProps } from '../runtime/jsx.ts'
 import { invariant } from '../runtime/invariant.ts'
-import { processStyleClass } from '../style/index.ts'
-import type { CSSProps } from '../style/style.ts'
+import { processStyleClass } from '../style/style.ts'
+import type { CSSProps } from '../style/properties.ts'
 
 type StyleEntry = { selector: string; css: string }
 type StyleCache = Map<string, StyleEntry>
