--- packages/ui/src/style/index.ts
@@ -1,9 +1,0 @@
-import { createStyleManager } from './stylesheet.ts'
-
-export type {
-  CSSProps as EnhancedStyleProperties,
-  DOMStyleProperties as StyleProperties,
-} from './style.ts'
-export { processStyleClass, normalizeCssValue } from './style.ts'
-export { createStyleManager }
-export type { ServerStyleSource, StyleManager } from './stylesheet.ts'
